Fixed template and added link to flint story
The template didn't support captions for "heading" images. I added the
code to support it, but it breaks for really long captions. Generally, I
don't like the look of the caption on the heading image right now-- it
looks better on the feature image. We should fix that at some point. As
for the Flint article itself, I added a link to the Google blog post
announcing the partnership.