Initial commits for landing-fixes
Changed _layouts/home.html so that the `home` layout supports a title
and content. At the moment, this is the quickest way to be able to
modify the landing page. If we want to do anything more specific, we can
always just modify the layout directly and add the content we want.
That's really messy, and I wouldn't advise it, but hey, 'move fast break
things', amirite?