Added a script and updated about page
The script is really messy, but it's super short so it doesn't really
matter. It takes the spreadsheet from google docs (assumed to be named
'sheet.csv' in the working dir--google doesn't have a drive API) and
outputs (to STDOUT) some markdown content of all of the people in that
spreadsheet. Makes it easier to keep the active roster up to date. All
you gotta do is put out a google form at the beginning of the semester,
run the script `python membership.py | pbcopy` (on mac) and suddenly the
markdown is on your clipboard.