Adding encrypted aws credentials
This is likely awful security practice, but for the time being it is
probably okay. This file is pgp encrypted and requires a password to
decrypt. For obvious reasons, please do not store the result in this
repo. The info in the file is required for the push-to-live script to
work.